Add appveyor.yml

+++ setup.py
@@ -38,22 +38,25 @@ class DistCommand(Command):
         import platform
 
         assert os.getcwd() == self.cwd, 'Must be in package root: %s' % self.cwd
-        os.system('rm -rf ./dist/syncrypt')
-        os.system('PYTHONPATH=lib/python3.5/site-packages/ pyinstaller syncrypt.spec')
-        #os.system('cp ./dist/syncrypt_gui/* ./dist/syncrypt')
         zipname = '{name}-{version}.{platform}-{machine}.zip'.format(
                 name=__name__,
                 version=__version__,
                 platform=platform.system().lower(),
                 machine=platform.machine()
         )
-        os.system('mkdir ./dist/syncrypt')
-        os.system('cp README.md LICENSE ./dist/syncrypt')
-        os.system('cp dist/syncrypt-bin ./dist/syncrypt/syncrypt')
-        os.system('cp dist/syncrypt_daemon ./dist/syncrypt/syncrypt_daemon')
-        os.system('cp dist-files/* ./dist/syncrypt/')
-        os.system('cd dist; rm -f {0}; zip {0} -r ./syncrypt'.format(zipname))
-        os.system('cd dist; shasum -a 256 {0} > {0}.sha256'.format(zipname))
+        if platform.system().lower() == 'windows':
+            os.system('pyinstaller scripts/syncrypt')
+            os.system('pyinstaller scripts/syncrypt_daemon')
+        else:
+            os.system('rm -rf ./dist/syncrypt')
+            os.system('PYTHONPATH=lib/python3.5/site-packages/ pyinstaller syncrypt.spec')
+            os.system('mkdir ./dist/syncrypt')
+            os.system('cp README.md LICENSE ./dist/syncrypt')
+            os.system('cp dist/syncrypt-bin ./dist/syncrypt/syncrypt')
+            os.system('cp dist/syncrypt_daemon ./dist/syncrypt/syncrypt_daemon')
+            os.system('cp dist-files/* ./dist/syncrypt/')
+            os.system('cd dist; rm -f {0}; zip {0} -r ./syncrypt'.format(zipname))
+            os.system('cd dist; shasum -a 256 {0} > {0}.sha256'.format(zipname))
         print("Generated {0}".format(os.path.join('dist', zipname)))
 
 cmdclass['dist'] = DistCommand
